+++ ostap/fitting/tests/test_fitting_resolutions2.py
@@ -432,7 +432,7 @@ def test_bukin2 () :
 
     logger.info ('Test Bukin2 : Gaussian with asymmetric exponential tails ' )
     from   ostap.fitting.resolution import ResoBukin2
-    reso = ResoBukin2 ( 'Bk2' , mass ,
+    reso = ResoBukin2 ( 'Bukin2' , mass ,
                         kA       = ( -1.e-3 , -1000  , -1.e-5 ) , 
                         kB       = ( +1.e-3 , +1.e-5 , +1000  ) , 
                         varsigma = ( 0.1 , 0.01 , 5.0 ) ,                               
@@ -553,6 +553,39 @@ def test_skewgenerror () :
     models.add ( reso)
 
 
+# =============================================================================
+## GenLogisticIV 
+# =============================================================================
+def test_genlogistic4 () :
+    
+    logger = getLogger ( 'test_genlogisic4' )
+
+    logger.info ('Test GenLogisticIV : Generalized Logistic Tyep IV' )
+    from   ostap.fitting.resolution import ResoGenLogisticIV
+    reso = ResoGenLogisticIV  ( 'GL4' , mass ,
+                                sigma  = ( 0.5 , 0.01 , 5.0 ) ,
+                                gamma  = ( 0.8 , 0.01 , 20  ) , 
+                                mean   = ( 0   , -1.5 , 1.5 ) , 
+                                kappa  = ( 0.1 , -1 , 1 ) )
+    
+    reso.kappa.fix() 
+    result, frame = reso. fitTo ( dataset , silent = True  )
+    
+    reso.kappa.release() 
+    result, frame = reso. fitTo ( dataset , silent = True  )
+
+    with wait ( 1 ) , use_canvas ( 'test_genlogistic4' ) : 
+        result, frame = reso. fitTo ( dataset , silent = True , draw = True )
+        
+    if 0 != result.status() or 3 != result.covQual() :
+        logger.warning('Fit is not perfect MIGRAD=%d QUAL=%d ' % ( result.status() , result.covQual () ) )
+        print(result)
+    else :     
+        make_print ( reso , result , 'GenLogisticIV', logger )
+ 
+    models.add ( reso)
+
+
 # ==============================================================================
 ## dump all models
 # ==============================================================================
@@ -643,19 +676,22 @@ if '__main__' == __name__ :
         test_hyperbolic  () ## Hyperbolic resolution model
     
     with timing ("GenHyperbolic" , logger ) :  
-        test_genhyperbolic  () ## generalised Hyperbolic resolution model
+        test_genhyperbolic  () ## Generalised Hyperbolic resolution model
 
     with timing ("Hypatia" , logger ) :  
         test_hypatia        () ## Hypatia resoltuion model
         
     with timing ("Das"      , logger ) :  
-        test_das           ()   ## Das resolution model
+        test_das           () ## Das resolution model
 
     with timing ("Bukin2"  , logger ) :  
-        test_bukin2      ()   ## Bukin2 resolution model
+        test_bukin2      ()  ## Bukin2 resolution model
 
     with timing ("Normal Laplace"  , logger ) :  
         test_normlapl      ()   ## Normal Laplace resolution model
+
+    with timing ("GenLogisticIV"  , logger ) :  
+        test_genlogistic4  ()   ## Generalized Logistic Type IV resolution model
 
     with timing ("PearsonIV"  , logger ) :  
         test_pearsonIV      ()   ## PearsonIV resolution model